requirements.txt
pandas==1.2.3
python-binance==0.7.9
pydantic==1.8.1
certifi==2020.12.5
influxdb-client==1.15.0
pydantic==1.8.1
pytz==2021.1
Rx==3.1.1
six==1.15.0
typing-extensions==3.7.4.3
urllib3==1.26.3
